Simplify getopt on Visual Studio using header-only wingetopt.h
